--- src/__init__.py
@@ -1,3 +1,0 @@
-from optimizer import create_optimizer
-from trainarguments import TrainArgument
-from trainer import Trainer
